pack spellchecker function into hint.py
