+++ rich/color.py
@@ -1,5 +1,7 @@
 import platform
 import re
+
+from .hint import levenshtein_distance, find_closest_words
 from colorsys import rgb_to_hls
 from enum import IntEnum
 from functools import lru_cache
@@ -16,50 +18,6 @@ if TYPE_CHECKING:  # pragma: no cover
 
 
 WINDOWS = platform.system() == "Windows"
-
-
-def levenshtein_distance(s1, s2):
-    if len(s1) < len(s2):
-        s1, s2 = s2, s1
-
-    # Create two lists of size len(s2) + 1
-    prev_row = list(range(len(s2) + 1))
-    curr_row = [0] * (len(s2) + 1)
-
-    for i, c1 in enumerate(s1):
-        curr_row[0] = i + 1
-
-        for j, c2 in enumerate(s2):
-            # If characters are the same, use the previous diagonal value
-            if c1 == c2:
-                curr_row[j + 1] = prev_row[j]
-            else:
-                # If characters are different, use the minimum of the adjacent cells + 1
-                curr_row[j + 1] = min(prev_row[j], prev_row[j + 1], curr_row[j]) + 1
-
-        # Copy the current row to the previous row for the next iteration
-        prev_row, curr_row = curr_row, prev_row
-
-    return prev_row[-1]
-
-# Uses levenshtein distance to find similarly spelt words
-# Returns string separated by 'or' if multiple matches found 
-def find_closest_words(user_word, correct_words):
-    min_distance = float("inf")
-    closest_words = []
-    for word in correct_words:
-        distance = levenshtein_distance(user_word, word)
-
-        # If a new minimum distance is found, clear the closest words list and update min_distance
-        if distance <= min_distance:
-            min_distance = distance
-            closest_words = [word]
-        elif distance == min_distance:
-            # If the current word has the same minimum distance, add it to the list
-            closest_words.append(word)
-    
-    if closest_words == []: return ''
-    return closest_words if len == 1 else ' or '.join(closest_words)
 
 
 
@@ -495,7 +453,7 @@ class Color(NamedTuple):
         color_match = RE_COLOR.match(color)
         if color_match is None:
             closest_color = find_closest_words(color,ANSI_COLOR_NAMES)
-            raise ColorParseError(f"{original_color!r} is not a valid color did you mean {closest_color}?")
+            raise ColorParseError(f"{original_color!r} is not a valid color, did you mean {closest_color}?")
 
         color_24, color_8, color_rgb = color_match.groups()
         if color_24:
