+++ rich/console.py
@@ -598,287 +598,6 @@ def detect_legacy_windows() -> bool:
     return WINDOWS and not get_windows_console_features().vt
 
 
-def levenshtein_distance(s1, s2):
-    if len(s1) < len(s2):
-        s1, s2 = s2, s1
-
-    # Create two lists of size len(s2) + 1
-    prev_row = list(range(len(s2) + 1))
-    curr_row = [0] * (len(s2) + 1)
-
-    for i, c1 in enumerate(s1):
-        curr_row[0] = i + 1
-
-        for j, c2 in enumerate(s2):
-            # If characters are the same, use the previous diagonal value
-            if c1 == c2:
-                curr_row[j + 1] = prev_row[j]
-            else:
-                # If characters are different, use the minimum of the adjacent cells + 1
-                curr_row[j + 1] = min(prev_row[j], prev_row[j + 1], curr_row[j]) + 1
-
-        # Copy the current row to the previous row for the next iteration
-        prev_row, curr_row = curr_row, prev_row
-
-    return prev_row[-1]
-stlye_set = {"black",
-   "red",
-    "green", 
-    "yellow",
-    "blue",
-    "magenta",
-    "cyan",
-    "microsoft",
-    "white",
-    "bright_black",
-    "bright_red",
-    "bright_green",
-    "bright_yellow",
-   "bright_blue",
-   "bright_magenta",
-   "bright_cyan",
-   "bright_white",
-   "grey0",
-   "gray0",
-   "navy_blue",
-   "dark_blue",
-   "blue3",
-   "blue1",
-   "dark_green",
-   "deep_sky_blue4",
-   "dodger_blue3",
-   "dodger_blue2",
-   "green4",
-   "spring_green4",
-   "turquoise4",
-   "deep_sky_blue3",
-   "dodger_blue1",
-   "green3",
-   "spring_green3",
-   "dark_cyan",
-   "light_sea_green",
-   "deep_sky_blue2",
-   "deep_sky_blue1",
-   "spring_green2",
-   "cyan3",
-   "dark_turquoise",
-   "turquoise2",
-   "green1",
-   "spring_green1",
-   "medium_spring_green",
-   "cyan2",
-   "cyan1",
-   "dark_red",
-   "deep_pink4",
-   "purple4",
-   "purple3",
-   "blue_violet",
-   "orange4",
-   "grey37",
-   "gray37",
-   "medium_purple4",
-   "slate_blue3",
-   "royal_blue1",
-   "chartreuse4",
-   "dark_sea_green4",
-   "pale_turquoise4",
-   "steel_blue",
-   "steel_blue3",
-   "cornflower_blue",
-   "chartreuse3",
-   "cadet_blue",
-   "sky_blue3",
-   "steel_blue1",
-   "pale_green3",
-   "sea_green3",
-   "aquamarine3",
-   "medium_turquoise",
-   "chartreuse2",
-   "sea_green2",
-   "sea_green1",
-   "aquamarine1",
-   "dark_slate_gray2",
-   "dark_magenta",
-   "dark_violet",
-   "purple",
-   "light_pink4",
-   "plum4",
-   "medium_purple3",
-   "slate_blue1",
-   "yellow4",
-   "wheat4",
-   "grey53",
-   "gray53",
-   "light_slate_grey",
-   "light_slate_gray",
-   "medium_purple",
-   "light_slate_blue",
-   "dark_olive_green3",
-   "dark_sea_green",
-   "light_sky_blue3",
-   "sky_blue2",
-   "dark_sea_green3",
-   "dark_slate_gray3",
-   "sky_blue1",
-   "chartreuse1",
-   "light_green",
-   "pale_green1",
-   "dark_slate_gray1",
-   "red3",
-   "medium_violet_red",
-   "magenta3",
-   "dark_orange3",
-   "indian_red",
-   "hot_pink3",
-   "medium_orchid3",
-   "medium_orchid",
-   "medium_purple2",
-   "dark_goldenrod",
-   "light_salmon3",
-   "rosy_brown",
-   "grey63",
-   "gray63",
-   "medium_purple1",
-   "gold3",
-   "dark_khaki",
-   "navajo_white3",
-   "grey69",
-   "gray69",
-   "light_steel_blue3",
-   "light_steel_blue",
-   "yellow3",
-   "dark_sea_green2",
-   "light_cyan3",
-   "light_sky_blue1",
-   "green_yellow",
-   "dark_olive_green2",
-   "dark_sea_green1",
-   "pale_turquoise1",
-   "deep_pink3",
-   "magenta2",
-   "hot_pink2",
-   "orchid",
-   "medium_orchid1",
-   "orange3",
-   "light_pink3",
-   "pink3",
-   "plum3",
-   "violet",
-   "light_goldenrod3",
-   "tan",
-   "misty_rose3",
-   "thistle3",
-   "plum2",
-   "khaki3",
-   "light_goldenrod2",
-   "light_yellow3",
-   "grey84",
-   "gray84",
-   "light_steel_blue1",
-   "yellow2",
-   "dark_olive_green1",
-   "honeydew2",
-   "light_cyan1",
-   "red1",
-   "deep_pink2",
-   "deep_pink1",
-   "magenta1",
-   "orange_red1",
-   "indian_red1",
-   "hot_pink",
-   "dark_orange",
-   "salmon1",
-   "light_coral",
-   "pale_violet_red1",
-   "orchid2",
-   "orchid1",
-   "orange1",
-   "sandy_brown",
-   "light_salmon1",
-   "light_pink1",
-   "pink1",
-   "plum1",
-   "gold1",
-   "navajo_white1",
-   "misty_rose1",
-   "thistle1",
-   "yellow1",
-   "light_goldenrod1",
-   "khaki1",
-   "wheat1",
-   "cornsilk1",
-   "grey100",
-   "gray100",
-   "grey3",
-   "gray3",
-   "grey7",
-   "gray7",
-   "grey11",
-   "gray11",
-   "grey15",
-   "gray15",
-   "grey19",
-   "gray19",
-   "grey23",
-   "gray23",
-   "grey27",
-   "gray27",
-   "grey30",
-   "gray30",
-   "grey35",
-   "gray35",
-   "grey39",
-   "gray39",
-   "grey42",
-   "gray42",
-   "grey46",
-   "gray46",
-   "grey50",
-   "gray50",
-   "grey54",
-   "gray54",
-   "grey58",
-   "gray58",
-   "grey62",
-   "gray62",
-   "grey66",
-   "gray66",
-   "grey70",
-   "gray70",
-   "grey74",
-   "gray74",
-   "grey78",
-   "gray78",
-   "grey82",
-   "gray82",
-   "grey85",
-   "gray85",
-   "grey89",
-   "gray89",
-   "grey93",
-   "gray93"}
-
-def find_closest_words(user_word, correct_words):
-    min_distance = float("inf")
-    closest_words = []
-    for word in correct_words:
-        distance = levenshtein_distance(user_word, word)
-
-        # If a new minimum distance is found, clear the closest words list and update min_distance
-        if distance <= min_distance:
-            min_distance = distance
-            closest_words = [word]
-        elif distance == min_distance:
-            # If the current word has the same minimum distance, add it to the list
-            closest_words.append(word)
-
-    return (closest_words, min_distance)
-
-# def find_closest_words(user_input, correct_words):
-#     res = []
-#     for word in user_input.split(' '):
-#         res.append(find_closest_word(word, correct_words))
-#     return res
 
 class Console:
     """A high level console interface.
@@ -1744,46 +1463,7 @@ class Console:
     
     
 
-    # def get_style(
-    #     self, name: Union[str, Style], *, default: Optional[Union[Style, str]] = None
-    # ) -> Style:
-    #     """Get a Style instance by its theme name or parse a definition.
-
-    #     Args:
-    #         name (str): The name of a style or a style definition.
-
-    #     Returns:
-    #         Style: A Style object.
-
-    #     Raises:
-    #         MissingStyle: If no style could be parsed from name.
-
-    #     """
-    #     if isinstance(name, Style):
-    #         return name
-
-    #     try:
-    #         style = self._theme_stack.get(name)
-    #         if style is None:
-    #             style = Style.parse(name)
-    #         return style.copy() if style.link else style
-    #     except errors.StyleSyntaxError as error:
-    #         print("in here")
-    #         if default is not None:
-    #             return self.get_style(default)
-            
-    #         #loads up default words from dictionary
-    #         # spell = SpellChecker()
-    #         # style_set = set(map(str.strip, open('C:/Users/julia/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0/LocalCache/local-packages/Python310/site-packages/rich/stylenames.txt')))
-    
-    #         # Get a list of `likely` options (not necessary, just wanted to see)
-    #         # print(spell.candidates(name))
-    #         # correctStyle = spell.correction(name)
-    #         (correctStyle, distance) = find_closest_words(name,stlye_set)
-
-    #         raise errors.MissingStyle(
-    #             f"Unable to get style {name!r}; {error}, did you mean '{correctStyle}', with distance '{distance}'?"
-    #         ) from None
+ 
 
     def _collect_renderables(
         self,
@@ -4402,16 +4082,11 @@ class Console:
             return name
 
         try:
-            print("a")
             style = self._theme_stack.get(name)
-            print("b")
             if style is None:
-                print("c")
                 style = Style.parse(name)
-            print("d")
             return style.copy() if style.link else style
         except errors.StyleSyntaxError as error:
-            print("in here")
             if default is not None:
                 return self.get_style(default)
             
@@ -4422,7 +4097,7 @@ class Console:
             # Get a list of `likely` options (not necessary, just wanted to see)
             # print(spell.candidates(name))
             # correctStyle = spell.correction(name)
-            (correctStyle, distance) = find_closest_words(name,stlye_set)
+            
 
             raise errors.MissingStyle(
                 f"Failed to get style {name!r}; {error}"
